+++ Rest.API/Services/IFibonacciService.cs
@@ -4,7 +4,7 @@ namespace Rest.API.Services
 {
     public interface IFibonacciService
     {
-        Int64 Fibonacci(int n);
+        Int64 Fibonacci(Int64 n);
     }
 
 }
